% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceRuStatsAPI.R
\name{PriceRuStatsAPI}
\alias{PriceRuStatsAPI}
\title{Get stats}
\usage{
PriceRuStatsAPI(login = NULL, password = NULL, token = NULL,
  date_start = NULL, date_end = NULL, shop_id = NULL, id = NULL)
}
\arguments{
\item{login}{Your account e-mail}

\item{password}{Your account password}

\item{token}{Your API token, you can get it from technical support}

\item{date_start}{Start date}

\item{date_end}{End date}

\item{shop_id}{Your shop id, see \code{\link{https://biz.price.ru/shops}}. See "Magazin ID: "}

\item{id}{Your price-list id, see \code{\link{https://biz.price.ru/shops}}. See "ID: ..... Price-list activen"}
}
\description{
Function to get stats
}
\examples{
PriceRuStatsAPI()
}
